All the bugs seem to be fixed inside calculate_line.
Made the input more friendly to xfer over to codingame.

Added a new function to `move the ground up` so that the pathing will not go so close to it.

Also, I am not terribly convinced I can follow through with my plan of `follow the line`. Am I pot committed?